Migrate fps-manual-exposure test from legacy to pytest
